PaymentTest: add configuration infrastructure
Add a configuration infrastructure for the PaymentTest project. Prior to
this, changes on this program due to testing purposes would be
committed as code changes, which could be confusing for tracking the
whole project's actual logic history.

Example config files have been added for Windows and Linux. They should
be renamed to "App.config" with a proper configuration to take effect.

Create Config.cs, a wrapper around ConfigurationManager to manage the
aforementioned configurations.

Bump the PaymentTest project version.