sqlite-net: handle 32bit and 64bit native libraries
Handle both 32bit and 64bit instruction sets for sqlite3 native
libraries on Windows and Linux-like operating systems. This is needed
for the PagarMe.Mpos project to support both architectures.

This is necessary because P/Invoke's DllImport does not allow for
variable library path names. This means we need to create a class for
each architecture (with different library names) and specify all the
entrypoints. An "ISqliteImport" interface has been created to unify
these classes.

The expected names for the libraries are, respectively, "sqlite3_32" and
"sqlite3_64".