Separate Config file to Generic project
Config files were a part of the `PaymentTest` project, which meant
everyone implementing the SDK would have to create a `Config.cs` file
with pretty much the same content. We isolated this to the project's
`App.{Platform}.config.example` and simplify this. We also had to move
the `Config.cs` to the `PagarMe.Generic` project. Finally, we added
references to `PagarMe.Generic` on the `PaymentTest` solution.