Expose initialize to let client configure device
This commit changes how we configure the Service with the customer data
(encryption key, port, baud rate). This was set on App.config and now we
expose a method so the customer can alter it from the client. With this
commit we stop using the `requestCode`, and we now altered how we call
`mposBridge.GetContext` passing the `request.ContextId`.

The `Baud Rate` used to be configured directly on the service, but now
it's being configured by a WebSocket message sent from the client. This
broke the `DeviceManager` and we had to alter it to not load the
`_baudRate` property on this class. We moved this to `IDevice` and
`SerialDevice` to configure the `Baud Rate` when the device opens the
connection to the server. We're now instantiating the `SerialPort` on
the Open method as well.