Adjust Setup/Server project dependencies
Add the necessary files from libc to
`PagarMe.Mpos.Bridge.Server/PagarMe.Mpos.Bridge.Server.csproj` and
`Setup/Setup.vdproj` to make the project work properly (wasn't working
on Linux).

Another detail that's committed with this change is the
`options.EnsureDefaults();` call. This function is called whenever
Options is instantiated and sets the correct file to store the SQLite
database (we added this to `MposWebsocketService` and
`PagarMe.Mpos.Bridge.Server/Program.cs`).