Configure NLog/WebSocket-Sharp logs, add more logs
This commit has more than just the logs: we removed the `requestCode`
from WebSocketBehaviour (because this is sent by the client now) and
changed PagarMe.Mpos.Bridge.Service/Program.cs `ServiceBase.Run` to
avoid using an array (since there's only one service).

That said, the commit has a lot of changes regarding how we log
information. We've created the class `LoggerExtension` to avoid
repetition, and we encapsulate the try-catch block on this class.
A bunch of other files were changed to use this new class. We also
created a configuration for the NLog project, that tecnically was
already used by our project but was lacking configuration and was not
available everywhere we need it. We also added logs for `OnOpen` and
`OnClose`, and added more data to logs generated by `OnMessage`.

We avoided using NLog on `Pagarme.Mpos` because it didn't already have
the NLog dependency. The  `DeviceManager` need to use the
`tryOnLogException` method, so we now pass it as a reference.