Add missing dependencies to run service at Linux
We've added a try-catch block on `MposWebsocketService` on both
`OnStart` and `OnStop` methods. And we're now logging the exceptions.
This was done during development and seems like a good practice, so
we're leaving it here. The end goal for the commit is actually what is
going on `Setup/Setup.vdproj` that's the one really including
dependencies needed to run the service on Linux servers.