Show message if no device is found
This commit only deals with the example project to use the Mpos. We
didn't account for the error that happens when the device requested by
the client wasn't listed on the server. We fixed this on the sample
project so that our customers can have a standard way of using the SDK.