Avoid error on ReadInternal at Ubuntu
The WebSocket protocol provides the `Sec-WebSocket-Extensions` header
that, if present, the value indicates the protocol-level extension(s)
the client wishes to speak. This was causing some errors on
Linux (Ubuntu) due to unknown reasons and, since we couldn't find any
security-related reason to use the protocol extensions as of yet, we're
setting the WebSocketSharper to ignore extensions.

Error Message: `Internal error (no progress possible) ReadInternal`