Fix DllImports for Linux
This SDK uses 3 external libs:
- mpos
- tms
- sqlite

On windows we need to know if it's 32 or 64 bits. We changed this and
the commit broke the behaviour in Linux. This commits fix the Linux
issue without compromising the Windows solution.

On Linux we use the same DLL for both 32 and 64 bits.

We now have 3 files that share a lot of code. Sometime in the future we
should evaluate options, such as the use of Reflection, to avoid having
multiple files with the same code.