Bifrost State Machine
The order of messages is relevant to our WebSocket server. Some methods
need to have a proper order to function. This commit changes the way we
treat these cases.

On the task `handleRequest` we now lock the context and validate if the
events orders make sense. If they don't we throw an error.