Stop Linux from cutting process by time
The server kills inactive connections by default, after 60s of
inactivity at Linux (Ubuntu). To avoid the frequent closing of the
connections, we set `KeepClean` to false. On Windows this wasn't a
problem, but on Ubuntu we had some errors (running on a limited resource
VM).

If this causes errors, we may evaluate to close connections based on a
different time window or some other criteria.