Fix problem when user cancels at mpos
When the user cancels the transaction at the mpos, a null reference
was being throw and it hides the real error code.

The commit changes the code to handle this NullReference, adds a new
type of transaction return (Canceled) and adds error code to
transaction result, to give right answer to the mpos consumer.